chore: move MultiComboBox to ComboBox directory
